requirements.txt
instapy_cli
watchdog
python-dateutil
